+++ test/Halcyon.Api.Tests/Features/Manage/GetProfile/GetProfileEndpointTests.cs
@@ -1,19 +1,17 @@
-﻿using Halcyon.Api.Data;
-using Halcyon.Api.Features.Manage.GetProfile;
+﻿using Halcyon.Api.Features.Manage.GetProfile;
 using Microsoft.AspNetCore.Mvc.Testing;
-using Microsoft.Extensions.DependencyInjection;
 using System.Net;
 using System.Net.Http.Json;
 
 namespace Halcyon.Api.Tests.Features.Manage.GetProfile;
 
-public class GetProfileEndpointTests : IClassFixture<CustomWebApplicationFactory<Program>>
+public class GetProfileEndpointTests : IClassFixture<TestWebApplicationFactory<Program>>
 {
     private const string RequestUri = "/manage";
 
     private readonly WebApplicationFactory<Program> factory;
 
-    public GetProfileEndpointTests(CustomWebApplicationFactory<Program> factory)
+    public GetProfileEndpointTests(TestWebApplicationFactory<Program> factory)
     {
         this.factory = factory;
     }
@@ -29,7 +27,7 @@ public class GetProfileEndpointTests : IClassFixture<CustomWebApplicationFactory
     [Fact]
     public async Task GetProfile_WhenAuthorized_ShouldReturnProfile()
     {
-        var user = await CreateTestUser();
+        var user = await factory.CreateTestUserAsync();
 
         var client = factory.CreateClient();
         client.DefaultRequestHeaders.Add(TestAuthenticationHandler.UserId, user.Id.ToString());
@@ -40,23 +38,5 @@ public class GetProfileEndpointTests : IClassFixture<CustomWebApplicationFactory
         var result = await response.Content.ReadFromJsonAsync<GetProfileResponse>();
         Assert.NotNull(result);
         Assert.Equal(user.Id, result.Id);
-    }
-
-    private async Task<User> CreateTestUser()
-    {
-        var user = new User
-        {
-            EmailAddress = $"{Guid.NewGuid()}@example.com",
-            FirstName = "Test",
-            LastName = "User",
-            DateOfBirth = new DateOnly(1070, 1, 1)
-        };
-
-        using var scope = factory.Services.CreateScope();
-        using var dbContext = scope.ServiceProvider.GetRequiredService<HalcyonDbContext>();
-        dbContext.Users.Add(user);
-        await dbContext.SaveChangesAsync();
-
-        return user;
     }
 }
